chore: moving incentivized tests into separate function
